server/requirements.txt
django==3.0.6
djangorestframework==3.11.0
django-cors-headers==3.2.1
django_extensions
mysqlclient==1.4.6
django-mysql==2.2.0
pymysql==0.9.3
sqlparse==0.3.1